Convert to TypeScript and ensure types are correct
